requirements.txt
selenium>=4.17.0
webdriver-manager>=4.0.1
googletrans==4.0.0-rc1
Pillow>=10.2.0
requests>=2.31.0
pytest>=8.0.0
beautifulsoup4>=4.12.0
